Relocating the test resources and providers to the src folder
