+++ src/groovy/org/grails/jaxrs/test/CustomRequestEntityReader.groovy
@@ -13,21 +13,22 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.jaxrs.itest
+package org.grails.jaxrs.test
 
-import javax.ws.rs.GET
-import javax.ws.rs.Path
-import javax.ws.rs.Produces
+import javax.ws.rs.Consumes
+import javax.ws.rs.core.MultivaluedMap
+import javax.ws.rs.ext.Provider
+
+import org.grails.jaxrs.support.MessageBodyReaderSupport
 
 /**
  * @author Martin Krasser
  */
-@Path('/test/01')
-class TestResource01 {
+@Provider
+@Consumes('text/plain')
+class CustomRequestEntityReader extends MessageBodyReaderSupport<CustomRequestEntity> {
 
-    @GET
-    @Produces('text/plain')
-    String test() {
-        'test01'
-    }
+     CustomRequestEntity readFrom(MultivaluedMap httpHeaders, InputStream entityStream) {
+         new CustomRequestEntity(content:entityStream.text)
+     }
 }
