+++ src/groovy/org/grails/jaxrs/test/TestPerson.groovy
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2009 - 2011 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.grails.jaxrs.test
+
+/**
+ * @author Martin Krasser
+ */
+class TestPerson {
+
+    // pretend being a domain class
+    def id
+    def version
+    void setClass(obj) {}
+
+    // test properties
+    String name
+    int age
+}
