+++ src/groovy/org/grails/jaxrs/test/TestResource02.groovy
@@ -1,5 +1,5 @@
 /*
- * Copyright 2009-2010 the original author or authors.
+ * Copyright 2009 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,21 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.jaxrs.itest
+package org.grails.jaxrs.test
 
-import javax.ws.rs.GET
+import org.grails.jaxrs.test.CustomRequestEntity
+import org.grails.jaxrs.test.CustomResponseEntity
+
+import javax.ws.rs.Consumes
+import javax.ws.rs.POST
 import javax.ws.rs.Path
 import javax.ws.rs.Produces
 
 /**
  * @author Martin Krasser
  */
-@Path('/test/05')
-class TestResource05 {
+@Path('/test/02')
+class TestResource02 {
 
-    @GET
-    @Produces('text/html')
-    String test() {
-        '<html><body>test05</body></html>'
+    @POST
+    @Consumes('text/plain')
+    @Produces('text/plain')
+    CustomResponseEntity test(CustomRequestEntity requestEntity) {
+        new CustomResponseEntity(content:'response:' + requestEntity.content)
     }
 }
