+++ src/groovy/org/grails/jaxrs/test/TestResource03.groovy
@@ -13,11 +13,28 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.grails.jaxrs.itest
+package org.grails.jaxrs.test
+
+import grails.converters.JSON
+
+import javax.ws.rs.Consumes
+import javax.ws.rs.POST
+import javax.ws.rs.Path
+import javax.ws.rs.Produces
 
 /**
  * @author Martin Krasser
  */
-class CustomResponseEntity {
-    String content
+@Path('/test/03')
+class TestResource03 {
+
+    @POST
+    @Consumes('application/json')
+    @Produces('application/json')
+    JSON testPerson(Map params) {
+        def person = new TestPerson(params)
+        person.name = person.name.reverse()
+        person.age = person.age + 1
+        return person as JSON
+    }
 }
