+++ src/groovy/org/grails/jaxrs/test/TestResource05.groovy
@@ -1,0 +1,33 @@
+/*
+ * Copyright 2009-2010 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.grails.jaxrs.test
+
+import javax.ws.rs.GET
+import javax.ws.rs.Path
+import javax.ws.rs.Produces
+
+/**
+ * @author Martin Krasser
+ */
+@Path('/test/05')
+class TestResource05 {
+
+    @GET
+    @Produces('text/html')
+    String test() {
+        '<html><body>test05</body></html>'
+    }
+}
