+++ test/integration/org/grails/jaxrs/itest/JaxrsControllerIntegrationTests.groovy
@@ -1,4 +1,14 @@
 package org.grails.jaxrs.itest
+
+import org.grails.jaxrs.test.CustomRequestEntityReader
+import org.grails.jaxrs.test.CustomResponseEntityWriter
+import org.grails.jaxrs.test.TestResource01
+import org.grails.jaxrs.test.TestResource02
+import org.grails.jaxrs.test.TestResource03
+import org.grails.jaxrs.test.TestResource04
+import org.grails.jaxrs.test.TestResource05
+import org.grails.jaxrs.test.TestResource06
+
 /*
  * Copyright 2009 - 2011 the original author or authors.
  *
